moved routes and training series files to new folder
