+++ src/components/VolunteerComponents/Pages/Dashboard/Routes/Routes.js
@@ -5,6 +5,7 @@ import Profile from "components/Pages/Profile";
 import EditTeamMember from "components/Pages/TeamMembers/Edit/";
 import CreateTrainingSeries from "components/Pages/TrainingSeries/Add/CreateTrainingSeries";
 import EditTrainingSeries from "components/Pages/TrainingSeries/Edit/";
+import SingleTrainingSeries from "components/Pages/TrainingSeries/SinglePage/";
 import AddTeamMember from "components/Pages/TeamMembers/Add/";
 import CreateMessage from "components/Pages/TrainingSeries/Add/CreateMessage";
 import MessagePage from "components/Pages/TrainingSeries/Add/MessagePage";
@@ -92,6 +93,17 @@ function Routes(props) {
         path="/home/training-series/:id/edit"
         render={renderProps => (
           <EditTrainingSeries {...renderProps} user_id={props.user.id} />
+        )}
+      />
+      <Route
+        path="/home/training-series/:id"
+        render={renderProps => (
+          <SingleTrainingSeries
+            {...renderProps}
+            props={props}
+            history={props.history}
+            activeTrainingSeries={props.activeTrainingSeries}
+          />
         )}
       />
       <Route
