+++ src/components/VolunteerComponents/Pages/Dashboard/Routes/index.js
@@ -1,0 +1,2 @@
+import Routes from "./Routes.js";
+export default Routes;
