+++ src/components/VolunteerComponents/Pages/TrainingSeries/SinglePage/index.js
