+++ src/components/VolunteerComponents/Pages/TrainingSeries/SinglePage/styles.js
